MDEV-23887: check_linker_flags - cmake compatibility
CMake-3.18 added a CHECK_LINKER_FLAGS that makes the check easy.

It was only from CMake-3.14.0 that CMAKE_REQUIRED_LINK_OPTIONS
was used in CHECK_CXX_SOURCE_COMPILES. Without this, it could be
the case (as was on OSX) that a flag was never checked in
CHECK_CXX_SOURCE_COMPILES, it successfully passed, but failed at link
time.

So prior to CMake-3.14.0, we use the compile flags check like we
used to because those got reused by the linker.